Change name to 160, adjust MCU multipliers of beta rev
