+++ keyboards/nazare160/beta/halconf.h
