+++ keyboards/nazare160/beta/mcuconf.h
@@ -25,9 +25,9 @@
 #undef STM32_PPRE2
 
 #define STM32_PLLM_VALUE                    4
-#define STM32_PLLN_VALUE                    72
+#define STM32_PLLN_VALUE                    96
 #define STM32_PLLP_VALUE                    2
-#define STM32_PLLQ_VALUE                    3
+#define STM32_PLLQ_VALUE                    4
 #define STM32_PPRE1                         STM32_PPRE1_DIV2
 #define STM32_PPRE2                         STM32_PPRE2_DIV1
